+++ include/audio_render_stage/audio_tape_render_stage.h
@@ -3,7 +3,8 @@
 #define AUDIO_TAPE_RENDER_STAGE_H
 
 #include "audio_core/audio_render_stage.h"
-#include "audio_render_stage/audio_render_stage_history.h"
+#include "audio_render_stage_plugins/audio_render_stage_history.h"
+#include "audio_core/audio_tape.h"
 
 typedef std::vector<float> Tape;
 class AudioRecordRenderStage : public AudioRenderStage {
