+++ lib/audio_render_stage/audio_tape_render_stage.cpp
@@ -5,7 +5,8 @@
 #include "audio_parameter/audio_texture2d_parameter.h"
 #include "audio_parameter/audio_uniform_parameter.h"
 #include "audio_render_stage/audio_tape_render_stage.h"
-#include "audio_render_stage/audio_render_stage_history.h"
+#include "audio_render_stage_plugins/audio_render_stage_history.h"
+#include "audio_core/audio_tape.h"
 
 const std::vector<std::string> AudioRecordRenderStage::default_frag_shader_imports = {
     "build/shaders/global_settings.glsl",
