+++ tests/audio_render_stage_history_gl_test.cpp
@@ -6,7 +6,8 @@
 #include "framework/test_main.h"
 
 #define private public
-#include "audio_render_stage/audio_render_stage_history.h"
+#include "audio_render_stage_plugins/audio_render_stage_history.h"
+#include "audio_core/audio_tape.h"
 #undef private
 
 #include "audio_core/audio_render_stage.h"
